+++ packages/miniapp-render/src/node/element.js
@@ -6,7 +6,7 @@ import ClassList from './class-list';
 import Style from './style';
 import Attribute from './attribute';
 import cache from '../utils/cache';
-import { toDash } from '../utils/tool';
+import { toDash, omitFalsyFields } from '../utils/tool';
 import { simplifyDomTree, traverse } from '../utils/tree';
 import { BUILTIN_COMPONENT_LIST, STATIC_COMPONENTS, PURE_COMPONENTS, CATCH_COMPONENTS, APPEAR_COMPONENT, ANCHOR_COMPONENT } from '../constants';
 
@@ -32,7 +32,6 @@ class Element extends Node {
     if (this.id && !this.ownerDocument.__idMap.has(this.id)) {
       this.ownerDocument.__idMap.set(this.id, this);
     }
-    this.__componentWrapperId = null; // ComponentWrapper NodeId which this belongs to
   }
 
   // Override the _destroy method of the parent class
@@ -55,6 +54,8 @@ class Element extends Node {
 
   _triggerUpdate(payload, immediate = true) {
     payload.nodeId = this.__nodeId;
+    payload.componentWrapperId = this.componentWrapperId;
+
     if (immediate) {
       this._enqueueRender(payload);
     } else {
@@ -118,6 +119,8 @@ class Element extends Node {
     if (hasCatchTouchMoveFlag) {
       CATCH_COMPONENTS.has(this.__tmplName) && (nodeTypePrefix = 'catch-');
     }
+
+    // Fix scroll-view shake problem caused by scroll-left or scroll-top
     if (isWeChatMiniProgram && hasAnchorScrollFlag) {
       ANCHOR_COMPONENT === this.__tmplName && (nodeTypePrefix = 'anchor-');
     }
@@ -127,14 +130,13 @@ class Element extends Node {
   get _renderInfo() {
     const nodeType = this._processNodeType();
 
-    return {
+    return omitFalsyFields({
       nodeType,
       nodeId: this.__nodeId,
-      pageId: this.__pageId,
       ...this.__attrs.__value,
       style: this.style.cssText,
       class: this.__isBuiltinComponent ? this.className : `h5-${this.__tagName} ${this.className}`,
-    };
+    }, ['class', 'style']);
   }
 
   get _internal() {
@@ -438,6 +440,7 @@ class Element extends Node {
     }
     // Set parentNode
     node.parentNode = this;
+
     if (this._isRendered()) {
       node.__rendered = true;
       // Trigger update