+++ packages/miniapp-render/src/node/element/custom-component.js
@@ -1,6 +1,7 @@
 import Element from '../element';
 import cache from '../../utils/cache';
-import { getId } from '../../utils/tool';
+import { getId, omitFalsyFields } from '../../utils/tool';
+import { NATIVE_TYPES } from '../../constants';
 
 class CustomComponent extends Element {
   constructor(options) {
@@ -15,20 +16,19 @@ class CustomComponent extends Element {
   }
 
   get _renderInfo() {
-    const renderInfo = {
+    const renderInfo = omitFalsyFields({
       nodeId: this.__nodeId,
-      pageId: this.__pageId,
       nodeType: this.__tagName,
       style: this.style.cssText,
-      className: this.className,
+      class: this.className,
       ...this.__attrs.__value
-    };
+    }, ['class', 'style']);
 
     const config = cache.getConfig();
     let nativeInfo = null;
-    if (this.__nativeType === 'customComponent') {
+    if (this.__nativeType === NATIVE_TYPES.customComponent) {
       nativeInfo = config.usingComponents[this.__tagName];
-    } else if (this.__nativeType === 'miniappPlugin') {
+    } else if (this.__nativeType === NATIVE_TYPES.miniappPlugin) {
       nativeInfo = config.usingPlugins[this.__tagName];
     }
     if (nativeInfo) {