+++ packages/miniapp-render/src/node/root.js
@@ -2,7 +2,7 @@ import Element from './element';
 import cache from '../utils/cache';
 import perf from '../utils/perf';
 import { isFunction } from '../utils/tool';
-import { COMPONENT_WRAPPER } from '../constants';
+import { NATIVE_TYPES } from '../constants';
 
 class RootElement extends Element {
   constructor(options) {
@@ -48,79 +48,13 @@ class RootElement extends Element {
     const { mainPackageName } = cache.getConfig();
     const window = cache.getWindow(mainPackageName);
 
-    if (internal.$batchedUpdates) {
-      let callback;
-      // there is no need to aggregate arrays if $batchedUpdate and $spliceData exist
-      internal.$batchedUpdates(() => {
-        this.__renderStacks.forEach((task, index) => {
-          if (index === this.__renderStacks.length - 1) {
-            callback = () => {
-              window._trigger('setDataFinished');
-              if (process.env.NODE_ENV === 'development') {
-                perf.stop('setData');
-              }
-              let fn;
-              while (fn = this.__renderCallbacks.pop()) {
-                fn();
-              }
-            };
-            internal.firstRenderCallback();
-          }
-          if (task.type === 'children') {
-            const spliceArgs = [task.start, task.deleteCount];
-            internal.$spliceData({
-              [task.path]: task.item ? spliceArgs.concat(task.item) : spliceArgs
-            }, callback);
-          } else {
-            internal.setData({
-              [task.path]: task.value
-            }, callback);
-          }
-        });
-      });
-    } else {
-      const renderObject = {};
-      const childrenValuePaths = [];
-      // let hasCustomWrapper = false;
-      // const customWrapperUpdate = [];
+    const { rootStacks, componentWrapperObject } = this._filterRenderStacks();
+    const componentWrapperCount = Object.keys(componentWrapperObject).length;
+    let count = rootStacks.length + componentWrapperCount;
 
-      this.__renderStacks.forEach(task => {
-        const path = task.path;
-        const node = cache.getNode(task.nodeId);
-        if (node && node.__nativeType === COMPONENT_WRAPPER) {
-          // hasCustomWrapper = true;
-          // customWrapperUpdate.push({
-          //   node: node._internal,
-          //   data: {
-          //     [path.replace(node._path, '')]: task.value
-          //   }
-          // });
-          node._internal.setData({
-            [path.replace(node._path, '')]: task.value()
-          }, () => {
-            console.log('component update');
-          });
-        } else {
-          if (task.type === 'children') {
-            childrenValuePaths.push(path);
-          }
-          renderObject[path] = task.value;
-        }
-      });
-      for (const path in renderObject) {
-        // If the whole father children path is set, then its children path can be deleted
-        childrenValuePaths.forEach(cp => {
-          if (path.includes(cp) && cp !== path) {
-            delete renderObject[path];
-          }
-        });
-        const value = renderObject[path];
-        if (isFunction(value)) {
-          renderObject[path] = value();
-        }
-      }
-      internal.firstRenderCallback(renderObject);
-      internal.setData(renderObject, () => {
+    const callback = () => {
+      count--;
+      if (!count) {
         window._trigger('setDataFinished');
         let fn;
         while (fn = this.__renderCallbacks.pop()) {
@@ -129,10 +63,109 @@ class RootElement extends Element {
         if (process.env.NODE_ENV === 'development') {
           perf.stop('setData');
         }
+      }
+    };
+
+    if (rootStacks.length > 0) {
+      if (internal.$batchedUpdates) {
+        internal.firstRenderCallback();
+        this._batchedUpdated(internal, rootStacks, callback);
+      } else {
+        const renderObject = Object.create(null);
+        rootStacks.forEach((task) => {
+          renderObject[task.path] = isFunction(task.value) ? task.value() : task.value;
+        });
+        internal.firstRenderCallback(renderObject);
+        internal.setData(renderObject, callback);
+      }
+    }
+
+    if (componentWrapperCount > 0) {
+      Object.values(componentWrapperObject).forEach(({ node, data }) => {
+        if (node._internal.$batchedUpdates) {
+          internal.firstRenderCallback();
+          this._batchedUpdated(node._internal, data, callback);
+        } else {
+          const renderObject = Object.create(null);
+          data.forEach((task) => {
+            renderObject[task.path] = isFunction(task.value) ? task.value() : task.value;
+          });
+          internal.firstRenderCallback(renderObject);
+          node._internal.setData(renderObject, callback);
+        }
       });
     }
 
     this.__renderStacks = [];
+  }
+
+  _filterRenderStacks() {
+    const rootStacks = [];
+    const componentWrapperObject = Object.create(null);
+
+    const renderObject = Object.create(null);
+    const childrenValuePaths = [];
+    this.__renderStacks.forEach(task => {
+      const path = task.path;
+      if (task.type === 'children') {
+        childrenValuePaths.push(path);
+      }
+      renderObject[path] = task;
+    });
+
+    for (const path in renderObject) {
+      // If the whole father children path is set, then its children path can be deleted
+      childrenValuePaths.forEach(cp => {
+        if (path.includes(cp) && cp !== path) {
+          delete renderObject[path];
+        }
+      });
+
+      const task = renderObject[path];
+      if (task) {
+        const value = isFunction(task.value) ? task.value() : task.value;
+        let componentWrapperNode = null;
+        if (task.componentWrapperId) {
+          componentWrapperNode = cache.getNode(task.componentWrapperId);
+        }
+        if (componentWrapperNode) {
+          if (!componentWrapperObject[task.componentWrapperId]) {
+            componentWrapperObject[task.componentWrapperId] = {
+              node: componentWrapperNode,
+              data: []
+            };
+          }
+          componentWrapperObject[task.componentWrapperId].data.push({
+            ...task,
+            path: 'r' + path.replace(componentWrapperNode._path, '')
+          });
+        } else {
+          rootStacks.push(task);
+        }
+      }
+    }
+
+    return {
+      rootStacks,
+      componentWrapperObject
+    };
+  }
+
+  _batchedUpdated(internal, stacks, callback) {
+    internal.$batchedUpdates(() => {
+      stacks.forEach((task, index) => {
+        if (task.type === 'children') {
+          const spliceArgs = [task.start, task.deleteCount];
+          internal.$spliceData({
+            [task.path]: task.item ? spliceArgs.concat(task.item) : spliceArgs
+          }, callback);
+        } else {
+          internal.setData({
+            [task.path]: task.value
+          }, callback);
+        }
+      });
+    });
   }
 }